Adds additional VIFs to domU (-e).
Uses Xen MAC prefix of 00:16:3e, then <ethN> as the next octet.  The
last two octets of MAC are the last two octets of the IP address.
Separates the Postfix and Dovecot start scripts (in case something
wants to run without Dovecot, but wants to be able to deliver
mail--even if just locally).