Allows MTU in ifconfig.X files.
/sbin/ifup, /sbin/ifdown, /lib/lsb/bridge, and /lib/lsb/ipv4-config
have been modified to allow the /sbin scripts to handle the PHY
link--including setting the per-link MTU--and let the downstream
/lib/lsb scripts handle service-specific configuration.

Downstream scripts are passed the ifcongig.X file (instead of just the
short name).  All service-specific variables need to be set explicitly
(e.g., BRIDGE_IF).

This entailed making ipv4-static a bit smarter (it will use an explicit
IPV4_IF if set; otherwise, will use the first PHYS).

The script bridge needed to be modified to use BRIDGE_IF, and to stop
trying to handle the PHY links.