Allows XLAPP to bootstrap itself.
LFS had an issue in 7.0 which didn't account for the upstream problem
with readlink(2).  Because LFS uses a newer/different kernel/glibc,
those two return EINVAL when path is an empty string.  POSIX requires
ENOENT, and the tests haven't been "relaxed" yet.  This makes that
change for test-readlink.h and test-areadlink.h.