Separates /etc/resolv.conf from NIC configuration, and adds the new /lib/services/* scripts.
The /lib/services/* scripts were always added, but it used to be added
after bridging was installed.  The MTU fix I introduced into ifup
required changes to all the scripts, and this broke the pre-bridge
configuration.  This just make sure those scripts are installed
*before* the first reboot.

Also adds 30sec to the watchdog (30 was cutting it too close), for a
total of 60sec before reboot.