Fixes 2013 bug with SCRIPT_GUARD, Part 2.
Script was not exiting properly: SCRIPT_GUARD would 'continue' when a
guard was found (correct), but would not exit properly (needs to check
for loop termination, but did not--wrong).

Adds a more robust loop-exit mechanism: checks if "current" script is
greater than "end" at the start of the loop, instead of equals-to in
multiple places.  DRY.